Create shopping_info.db (Randomly generated when run)
I added this. It randomly generated when i ran the whole Riki but this db file is the one that actually saves user content. It may auto generate a new one if you don't include this. In Pycharm it will create an instance folder if you don't include this file.